Make Deimos optional and disabled by default
Summary:
Deimos+Mesos is not production-ready yet.

With Deimos 0.4.0 and Mesos 0.19, failed tasks become stuck in STAGING and are not reaped until mesos-slave is restarted. See:
https://github.com/mesosphere/deimos/issues/21
https://issues.apache.org/jira/browse/MESOS-1462

The Mesos fix probably won't land until 0.19.2 (if one is released) or 0.20.0.

Test Plan:
Launch a stack with Deimos enabled
Launch a stack with Deimos disabled

Reviewers: merkkila

Reviewed By: merkkila

Subscribers: tberman, justin, stephan, serdar

Differential Revision: http://phabricator.thefactory.com/D2825